Change imports to absolute imports
The scripts in the src/SigasiProjectCreator package are not compatible with explicit relative imports. This is because relative imports use the __name__ attribute to resolve modules. The __name__ attribute of a module which is run as a main script is set to "__main__", this breaks the relative imports used in the modules imported by the main script.
Another issue arose when updating the imports; the package contained a module with the same name (SigasiProjectCreator). The syntax for importing packages and modules in Python is similar, the clashing names caused issues in the imports. To fix this, the SigasiProjectCreator.py was renamed to Creator.py.