Make tests compatible with Python 3 (#20)
* Convert implicit relative imports to explicit relative imports

* Update tests to be compatible with both Python 2 and Python 3

Conversions include:
  - dict() key iteration: iteritems() in py2 -> items() in both py2 and py3
  - read() return val: str in py2, byte str in py3 -> explicit decode in both py2 and py3
  - Exception __contains__: comparison to repr / str in py2, not existent in py3 -> explicit str conversion in both py2 and py3

* Change imports to absolute imports

The scripts in the src/SigasiProjectCreator package are not compatible with explicit relative imports. This is because relative imports use the __name__ attribute to resolve modules. The __name__ attribute of a module which is run as a main script is set to "__main__", this breaks the relative imports used in the modules imported by the main script.
Another issue arose when updating the imports; the package contained a module with the same name (SigasiProjectCreator). The syntax for importing packages and modules in Python is similar, the clashing names caused issues in the imports. To fix this, the SigasiProjectCreator.py was renamed to Creator.py.

* Make tests work on Windows

Some test were failing on Windows because of issues with newlines (fixed by binary writing mode + encode) and because of illegal directory removal (fixed by restoring working directory after test).

* Fix invalid escape sequence warning

* Add Python 3.8 to .travis.yml and add note in README about compatible versions

* Use ConfigParser instead of deprecated SafeConfigParser